Bresenham is Bresenham, and what we do is what we do
