+++ core/src/test/java/com/graphhopper/storage/index/BresenhamLineTest.java
@@ -42,21 +42,21 @@ public class BresenhamLineTest {
 
     @Test
     public void testBresenhamLineLeftDown() {
-        BresenhamLine.calcPoints(5, 2, 0, 0, points::add);
+        BresenhamLine.bresenham(5, 2, 0, 0, points::add);
         // 5,2, 4,2, 3,2, 3,1, 2,1, 1,1, 0,0
         assertEquals(Helper.createPointList(5, 2, 4, 2, 3, 1, 2, 1, 1, 0, 0, 0), points);
     }
 
     @Test
     public void testBresenhamLineRightDown() {
-        BresenhamLine.calcPoints(3, 1, 0, 3, points::add);
+        BresenhamLine.bresenham(3, 1, 0, 3, points::add);
         // 3,1, 2,1, 1,1, 1,2, 0,2, 0,3
         assertEquals(Helper.createPointList(3, 1, 2, 2, 1, 2, 0, 3), points);
     }
 
     @Test
     public void testBresenhamLineLeftUp() {
-        BresenhamLine.calcPoints(2, 2, 3, 0, points::add);
+        BresenhamLine.bresenham(2, 2, 3, 0, points::add);
         // 2,2, 2,1, 2,0, 3,0
 
         assertEquals(Helper.createPointList(2, 2, 2, 1, 3, 0), points);
@@ -64,63 +64,9 @@ public class BresenhamLineTest {
 
     @Test
     public void testBresenhamLineRightUp() {
-        BresenhamLine.calcPoints(0, 0, 2, 3, points::add);
+        BresenhamLine.bresenham(0, 0, 2, 3, points::add);
         // 0,0, 0,1, 1,1, 1,2, 2,2, 2,3
         assertEquals(Helper.createPointList(0, 0, 1, 1, 1, 2, 2, 3), points);
     }
 
-    @Test
-    public void testBresenhamBug() {
-        BresenhamLine.calcPoints(0.5, -0.5, -0.6, 1.6, -1, -1, 0.75, 1.3, (lat, lon) -> points.add(lat, lon));
-        assertEquals(Helper.createPointList(0.575, -0.87, -0.175, 0.43, -0.925, 1.73), points);
-    }
-
-    @Test
-    public void testBresenhamHorizontal() {
-        BresenhamLine.calcPoints(.5, -.5, .5, 1, -1, -1, 0.6, 0.4, (lat, lon) -> points.add(lat, lon));
-        assertEquals(Helper.createPointList(.26, -.56, .26, -0.16, .26, .24, .26, .64, .26, 1.04), points);
-    }
-
-    @Test
-    public void testBresenhamVertical() {
-        BresenhamLine.calcPoints(-.5, .5, 1, 0.5, 0, 0, 0.4, 0.6, (lat, lon) -> points.add(lat, lon));
-        assertEquals(Helper.createPointList(-0.36, .06, 0.04, 0.06, 0.44, 0.06, 0.84, 0.06), points);
-    }
-
-    @Test
-    public void testRealBresenham() {
-        int parts = 4;
-        int bits = (int) (Math.log(parts * parts) / Math.log(2));
-        double minLon = -1, maxLon = 1.6;
-        double minLat = -1, maxLat = 0.5;
-        final KeyAlgo keyAlgo = new SpatialKeyAlgo(bits).setBounds(minLon, maxLon, minLat, maxLat);
-        double deltaLat = (maxLat - minLat) / parts;
-        double deltaLon = (maxLon - minLon) / parts;
-        final ArrayList<Long> keys = new ArrayList<>();
-        keys.clear();
-        BresenhamLine.calcPoints(.3, -.3, -0.2, 0.2, minLat, minLon, deltaLat, deltaLon, (lat1, lon1) -> keys.add(keyAlgo.encode(lat1, lon1))
-        );
-        assertEquals(Arrays.asList(11L, 9L), keys);
-
-        keys.clear();
-        BresenhamLine.calcPoints(.3, -.1, -0.2, 0.4, minLat, minLon, deltaLat, deltaLon, (lat, lon) -> keys.add(keyAlgo.encode(lat, lon))
-        );
-
-        // 11, 9, 12
-        assertEquals(Arrays.asList(11L, 12L), keys);
-
-        keys.clear();
-        BresenhamLine.calcPoints(.5, -.5, -0.1, 0.9, minLat, minLon, deltaLat, deltaLon, (lat, lon) -> keys.add(keyAlgo.encode(lat, lon))
-        );
-        // precise: 10, 11, 14, 12
-        assertEquals(Arrays.asList(10L, 11L, 12L), keys);
-    }
-
-    @Test
-    public void testBresenhamToLeft() {
-        BresenhamLine.calcPoints(
-                47.57383, 9.61984,
-                47.57382, 9.61890, 47, 9, 0.00647, 0.00964, points::add);
-        assertEquals(points.toString(), 1, points.getSize());
-    }
 }
